requirements-dev.txt
protobuf==3.19.4
grpcio==1.48.1
grpcio-reflection==1.48.1

accelbyte-py-sdk
bitarray
httpx
mmh3
pyjwt[crypto]
pyyaml
requests
websockets

grpcio-testing==1.48.1
grpcio-tools==1.48.1

black
pylint
tap.py
